Move candidates and matchers into submodule
