+++ fonduer/candidates/__init__.py
@@ -1,0 +1,8 @@
+from fonduer.candidates.candidates import (
+    CandidateExtractor,
+    Ngrams,
+    OmniFigures,
+    OmniNgrams,
+)
+
+__all__ = ["CandidateExtractor", "OmniFigures", "OmniNgrams", "Ngrams"]
